+++ statistics-impl/src/main/java/com/dwarfeng/statistics/impl/handler/provider/mock/RequiredBridgeDataValueType.java
@@ -1,0 +1,11 @@
+package com.dwarfeng.statistics.impl.handler.provider.mock;
+
+import java.lang.annotation.*;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.METHOD})
+@Documented
+@interface RequiredBridgeDataValueType {
+
+    String value();
+}