+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 #### 功能构建
 
+- 实现预设提供器。
+  - com.dwarfeng.statistics.impl.handler.provider.mock.MockProviderRegistry。
+
 - 实现预设映射器。
   - com.dwarfeng.statistics.impl.handler.mapper.WindowMapperRegistry。
   - com.dwarfeng.statistics.impl.handler.mapper.HighPassCounterMapperRegistry。